Revert "[BACKLOG-43558] - Webttle - Refactoring Textfile & CsvInput GetFields - 2"
