+++ engine/src/main/java/org/pentaho/di/trans/steps/util/CsvInputAwareStepUtil.java
@@ -11,7 +11,7 @@
  ******************************************************************************/
 
 
-package org.pentaho.di.trans.steps.common;
+package org.pentaho.di.trans.steps.util;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -28,6 +28,7 @@ import org.pentaho.di.core.logging.LogChannelInterface;
 import org.pentaho.di.core.util.Utils;
 import org.pentaho.di.i18n.BaseMessages;
 import org.pentaho.di.trans.TransMeta;
+import org.pentaho.di.trans.steps.common.CsvInputAwareMeta;
 import org.pentaho.di.trans.steps.csvinput.CsvInput;
 import org.pentaho.di.trans.steps.fileinput.text.BufferedInputStreamReader;
 import org.pentaho.di.trans.steps.fileinput.text.EncodingType;
@@ -35,7 +36,7 @@ import org.pentaho.di.trans.steps.fileinput.text.TextFileInputFieldDTO;
 import org.pentaho.di.trans.steps.fileinput.text.TextFileInputMeta;
 import org.pentaho.di.trans.steps.fileinput.text.TextFileInputUtils;
 
-public interface CsvInputAwareStep {
+public interface CsvInputAwareStepUtil {
 
   /**
    * Retrieves the field names from the CSV input metadata.
